Update data preprocessing module to fix some bugs
some changes in input checking so that the output of each function in the module is compatible with the following ones on the line.

make historical:

some refactoring.

re implement spatial historical part.

naming column names so that it is compatible with the prediction module of stpredict package

set spatial id:

creating a folder to save the plots

set temporal id:

add a new time format

deleted the verbose.

remove some unececery warnings.